Change suffix from gst to /gst
Adding a slash to the suffix makes it both easier to see and easier to search for.

Additionally, the next version of the "Show company branches" mod will use a single slash. If both mods use the exact same suffix, their load order relative to each other doesn't matter, which makes them easier to use.