Fix compatibility with ATS earlier than 1.46
With version 1 of this mod, the company logos failed to show up on the World Map in ATS 1.43. Apparently the "fueltank" body type was only introduced in ATS 1.45.

Adding "foodtank" as an alternative fixes this issue for 1.43. The "_fueldiel"/"_fuel_ldiel" body types may help in even earlier versions, which are untested.